Big improvements
Moved to a registry service architecture following this style:
https://rauljordan.com/2020/03/10/building-a-service-registry-in-go.html.

MongoJobsStorage now returns a channel used to get and save jobs, and it
fetches multiple jobs at a time from the database when necessary.

MongoJobsStorage and ElasticPageStorage now connects to the backend each
time it's necessary instead of mantaining one active connection the
whole time.

Fixed some race conditions on waitgroups and on Start functions